Move code to trait
